# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Waifu.im",
        "termsOfService": "https://www.waifu.im/terms-of-service",
        "contact": {
            "name": "Contact",
            "url": "http://www.waifu.im/contact",
            "email": "contact@waifu.im"
        },
        "license": {
            "name": "MPL 2.0",
            "url": "https://www.mozilla.org/en-US/MPL/2.0/"
        },
        "version": "5.2.2"
    },
    "host": "api.waifu.im",
    "basePath": "/",
    "paths": {
        "/fav": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves images based on the user's favorites.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "Fetch Favorite Images",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Force the API to return images with at least all the provided tags",
                        "name": "included_tags",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Force the API to return images without any of the provided tags",
                        "name": "excluded_tags",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": "58e6f0372364abda",
                        "description": "Force the API to provide only the specified file IDs or signatures",
                        "name": "included_files",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": "8108",
                        "description": "Force the API to not list the specified file IDs or signatures",
                        "name": "excluded_files",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "null",
                            "true",
                            "false"
                        ],
                        "type": "string",
                        "default": "false",
                        "description": "Force or exclude lewd files (only works if included_tags only contain versatile tags and no nsfw only tag).  You  can  provide  'null'  to  make  it  be  random.",
                        "name": "is_nsfw",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Force or prevent the API to return .gif files",
                        "name": "gif",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "FAVORITES",
                            "UPLOADED_AT",
                            "LIKED_AT",
                            "RANDOM"
                        ],
                        "type": "string",
                        "default": "RANDOM",
                        "description": "Ordering criteria",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "LANDSCAPE",
                            "PORTRAIT",
                            "RANDOM"
                        ],
                        "type": "string",
                        "default": "RANDOM",
                        "description": "Image orientation",
                        "name": "orientation",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Return an array of 30 files if true",
                        "name": "many",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Returns the full result without any limit (admins only)",
                        "name": "full",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003e=2000",
                        "description": "Filter images by width (in pixels).                                   Accepted  operators:  \u003c=,  \u003e=,  \u003e,  \u003c,  !=,  =",
                        "name": "width",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003e=2000",
                        "description": "Filter images by height (in pixels).                                  Accepted  operators:  \u003c=,  \u003e=,  \u003e,  \u003c,  !=,  =",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003e=2000",
                        "description": "Filter images by byte size. Accepted operators: \u003c=, \u003e=, \u003e, \u003c, !=, =",
                        "name": "byte_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.ImagesJsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/serializers.JSONError"
                        }
                    }
                }
            }
        },
...
        "/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves images randomly or by tag based on the specified search criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Images"
                ],
                "summary": "Search Images",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Force the API to return images with at least all the provided tags",
                        "name": "included_tags",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Force the API to return images without any of the provided tags",
                        "name": "excluded_tags",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": "58e6f0372364abda",
                        "description": "Force the API to provide only the specified file IDs or signatures",
                        "name": "included_files",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": "8108",
                        "description": "Force the API to not list the specified file IDs or signatures",
                        "name": "excluded_files",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "null",
                            "true",
                            "false"
                        ],
                        "type": "string",
                        "default": "false",
                        "description": "Force or exclude lewd files (only works if included_tags only contain versatile tags and no nsfw only tag).  You  can  provide  'null'  to  make  it  be  random.",
                        "name": "is_nsfw",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Force or prevent the API to return .gif files",
                        "name": "gif",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "FAVORITES",
                            "UPLOADED_AT",
                            "LIKED_AT",
                            "RANDOM"
                        ],
                        "type": "string",
                        "default": "RANDOM",
                        "description": "Ordering criteria",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "LANDSCAPE",
                            "PORTRAIT",
                            "RANDOM"
                        ],
                        "type": "string",
                        "default": "RANDOM",
                        "description": "Image orientation",
                        "name": "orientation",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Return an array of 30 files if true",
                        "name": "many",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Returns the full result without any limit (admins only)",
                        "name": "full",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003e=2000",
                        "description": "Filter images by width (in pixels).                                   Accepted  operators:  \u003c=,  \u003e=,  \u003e,  \u003c,  !=,  =",
                        "name": "width",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003e=2000",
                        "description": "Filter images by height (in pixels).                                  Accepted  operators:  \u003c=,  \u003e=,  \u003e,  \u003c,  !=,  =",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003e=2000",
                        "description": "Filter images by byte size. Accepted operators: \u003c=, \u003e=, \u003e, \u003c, !=, =",
                        "name": "byte_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.ImagesJsonResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/serializers.JSONError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Artist": {
            "type": "object",
            "properties": {
                "artist_id": {
                    "type": "integer",
                    "example": 1
                },
                "deviant_art": {
                    "type": "string",
                    "example": "https://www.deviantart.com/4thwallzart"
                },
                "name": {
                    "type": "string",
                    "example": "fourthwallzart"
                },
                "patreon": {
                    "type": "string"
                },
                "pixiv": {
                    "type": "string"
                },
                "twitter": {
                    "type": "string",
                    "example": "https://twitter.com/4thWallzArt"
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "artist": {
                    "$ref": "#/definitions/models.Artist"
                },
                "byte_size": {
                    "type": "integer",
                    "example": 3299586
                },
                "dominant_color": {
                    "type": "string",
                    "example": "#bbb7b2"
                },
                "extension": {
                    "type": "string",
                    "example": ".png"
                },
                "favorites": {
                    "type": "integer",
                    "example": 1
                },
                "height": {
                    "type": "integer",
                    "example": 2304
                },
                "image_id": {
                    "type": "integer",
                    "example": 8108
                },
                "is_nsfw": {
                    "type": "boolean",
                    "example": false
                },
                "liked_at": {
                    "type": "string"
                },
                "preview_url": {
                    "type": "string",
                    "example": "https://www.waifu.im/preview/8108/"
                },
                "signature": {
                    "type": "string",
                    "example": "58e6f0372364abda"
                },
                "source": {
                    "type": "string",
                    "example": "https://www.patreon.com/posts/persephone-78224476"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "uploaded_at": {
                    "type": "string",
                    "example": "2023-05-03T18:40:04.381354+02:00"
                },
                "url": {
                    "type": "string",
                    "example": "https://cdn.waifu.im/8108.png"
                },
                "width": {
                    "type": "integer",
                    "example": 1536
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A female anime/manga character."
                },
                "is_nsfw": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "waifu"
                },
                "tag_id": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "serializers.ImagesJsonResponse": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Image"
                    }
                }
            }
        },
        "serializers.JSONError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}